requirements.txt
absl-py==0.4.1
astor==0.7.1
audioread==2.1.6
cffi==1.11.5
Click==7.0
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
gast==0.2.0
grpcio==1.14.2
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.12.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
librosa==0.6.2
numpy==1.15.1
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.1
pycparser==2.19
pydot==1.4.1
PySoundFile==0.9.0.post1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.13
resampy==0.2.1
scikit-learn==0.19.2
scipy==1.0.0
seaborn==0.9.0
six==1.11.0
tensorflow==1.10.0
termcolor==1.1.0
webrtcvad==2.0.10
Werkzeug==0.14.1

